creazione Vista Vendite

+++ controllers/back/ajax/CDictionaryRemasterVideoAjaxController.php
@@ -147,7 +147,7 @@ class CDictionaryRemasterVideoAjaxController extends AAjaxController
                             $filenametoextrat=preg_replace('(.)', '_', $filenametoextrat, 1);
 
                         }
-                           ftp_get($conn_id,$remotetoLocalDirectory . '/' . $filenametoextrat,$image,FTP_BINARY);
+                        ftp_get($conn_id,$remotetoLocalDirectory . '/' . $filenametoextrat,$image,FTP_BINARY);
 
                         /*$file = fopen($remotetoLocalDirectory . '/' . $filenametoextrat, 'w');
                         curl_setopt($curl, CURLOPT_FILE, $file); #output
@@ -191,33 +191,25 @@ class CDictionaryRemasterVideoAjaxController extends AAjaxController
 
 
 
-                        ftp_put($conn_id, $remotepathOriginal . $directoryName . '_' . $resultdate . '/' . $filenametoextrat, $source, FTP_BINARY);
-
-/*$save_to_dir = '/home/iwespro/public_html/temp-remaster';
-                    $ftp_server = 'fiber.office.iwes.it';
-                    $pathlocal = '/home/iwespro/public_html/temp-remaster/';
-                    $save_to = '/home/iwespro/public_html/temp-remaster/';
-                    $save_to_dir = '/home/iwespro/public_html/temp-remaster';
-                    $path = 'shootImport/workvideo';
-                    $remotepathTodo = 'shootImport/newage2/topublish/';
-                    $remotepathOriginal = '/shootImport/newage2/original/';
-                    $remotepathToRename = '/shootImport/newage2/torename/';*/
+                    ftp_put($conn_id, $remotepathOriginal . $directoryName . '_' . $resultdate . '/' . $filenametoextrat, $source, FTP_BINARY);
 
 
 
-                        $filenameremaster = $save_to_dir . $item . '/' . $resultdate . '/' . 'remaster_'.$imagetoWorkName;
+
+
+                    $filenameremaster = $save_to_dir . $item . '/' . $resultdate . '/' . 'remaster_'.$imagetoWorkName;
 
 
 
                     $cmd = "ffmpeg -i ".$source." -vcodec copy -an ".$filenameremaster;
 //$cmd = "ffmpeg -y -i /media/sf_sites/iwespro/temp/video.mp4 -i /media/sf_sites/iwespro/temp/audio.mp3 -shortest -vcodec libx264 -acodec libfaac -b:v 1000k -refs 6 -coder 1 -sc_threshold 40 -flags +loop -me_range 16 -subq 7 -i_qfactor 0.71 -qcomp 0.6 -qdiff 4 -trellis 1 -b:a 128k -pass 1 -passlogfile /media/sf_sites/iwespro/temp-remaster/shootImport/resize/carte1610ok/2020-10-23/video3.mp4";
-                    exec($cmd,$output);
-                    sleep(4);
-                        ftp_put($conn_id, $remotepathTodo . $directoryName . '_' . $resultdate . '/' . $imagetoWorkName, $filenameremaster, FTP_BINARY);
-                        unlink($filenameremaster);
-                        unlink($save_to_dir . $item . '/' . $resultdate . '/' . $imagetoWorkName);
+                  if(exec($cmd,$output)) {
 
-                        // unlink($filenameremaster);
+                      ftp_put($conn_id,$remotepathTodo . $directoryName . '_' . $resultdate . '/' . $imagetoWorkName,$filenameremaster,FTP_BINARY);
+                      unlink($filenameremaster);
+                      unlink($save_to_dir . $item . '/' . $resultdate . '/' . $imagetoWorkName);
+                  }
+                    // unlink($filenameremaster);
